Refactor client side
